+++ ci/clean-storybook-files.js
@@ -1,17 +1,19 @@
 const fs = require('fs');
 const path = require('path');
 const allowedExtensions =
-  'css, gitignore, gitkeep, html, ico, jpg, js, json, png, scss, stackblitzrc, svg, ts'
+  'css, gitignore, gitkeep, html, ico, jpg, js, json, png, scss, stackblitzrc, svg, ts, txt, map'
     .split(/[, ]+/g)
     .map((e) => `.${e}`);
 
 // Removes all files not matching allowed extensions from given directory.
 function clean(dir) {
   fs.readdirSync(dir, { withFileTypes: true }).forEach((d) => {
+    const pathName = path.join(dir, d.name);
     if (d.isDirectory()) {
-      clean(path.join(dir, d.name));
-    } else if (d.isFile() && allowedExtensions.includes(path.extname(d.name))) {
-      fs.unlinkSync(path.join(dir, d.name));
+      clean(pathName);
+    } else if (d.isFile() && !allowedExtensions.includes(path.extname(d.name))) {
+      console.log(`Removing ${pathName}`);
+      fs.unlinkSync(pathName);
     }
   });
 }